requirements.txt
pytest==7.3.1
requests==2.31.0
jsonschema==4.19.0
curlify==2.2.1
allure-pytest==2.13.2
